print tables for pbpb / pp ratios. Now there are 16 tables and all the data is covered.

+++ ShellScripts/Data/photonJetFF/runPrintHEPDataHeaders.sh
@@ -12,6 +12,14 @@ $outputBase"/data5.yaml"
 $outputBase"/data6.yaml"
 $outputBase"/data7.yaml"
 $outputBase"/data8.yaml"
+$outputBase"/data9.yaml"
+$outputBase"/data10.yaml"
+$outputBase"/data11.yaml"
+$outputBase"/data12.yaml"
+$outputBase"/data13.yaml"
+$outputBase"/data14.yaml"
+$outputBase"/data15.yaml"
+$outputBase"/data16.yaml"
 );
 
 oldLine1="  values:"
@@ -20,6 +28,14 @@ newLines1=(
 "- header: {name: XI_(JET)}"
 "- header: {name: XI_(JET)}"
 "- header: {name: XI_(JET)}"
+"- header: {name: XI_(JET)}"
+"- header: {name: XI_(JET)}"
+"- header: {name: XI_(JET)}"
+"- header: {name: XI_(JET)}"
+"- header: {name: XI_T_(GAMMA)}"
+"- header: {name: XI_T_(GAMMA)}"
+"- header: {name: XI_T_(GAMMA)}"
+"- header: {name: XI_T_(GAMMA)}"
 "- header: {name: XI_T_(GAMMA)}"
 "- header: {name: XI_T_(GAMMA)}"
 "- header: {name: XI_T_(GAMMA)}"
@@ -28,14 +44,22 @@ newLines1=(
 
 oldLine2="  values:"
 newLines2=(
-"- header: {name: 1/N(JET) * DN(TRK))/DXI_(JET) PP}"
-"- header: {name: 1/N(JET) * DN(TRK))/DXI_(JET) PP}"
-"- header: {name: 1/N(JET) * DN(TRK))/DXI_(JET) PP}"
-"- header: {name: 1/N(JET) * DN(TRK))/DXI_(JET) PP}"
-"- header: {name: 1/N(JET) * DN(TRK))/XI_T_(GAMMA) PP}"
-"- header: {name: 1/N(JET) * DN(TRK))/XI_T_(GAMMA) PP}"
-"- header: {name: 1/N(JET) * DN(TRK))/XI_T_(GAMMA) PP}"
-"- header: {name: 1/N(JET) * DN(TRK))/XI_T_(GAMMA) PP}"
+"- header: {name: 1/N(JET) * DN(TRK))/DXI_(JET) PP smeared}"
+"- header: {name: 1/N(JET) * DN(TRK))/DXI_(JET) PP smeared}"
+"- header: {name: 1/N(JET) * DN(TRK))/DXI_(JET) PP smeared}"
+"- header: {name: 1/N(JET) * DN(TRK))/DXI_(JET) PP smeared}"
+"- header: {name: PBPB / PP}"
+"- header: {name: PBPB / PP}"
+"- header: {name: PBPB / PP}"
+"- header: {name: PBPB / PP}"
+"- header: {name: 1/N(JET) * DN(TRK))/XI_T_(GAMMA) PP smeared}"
+"- header: {name: 1/N(JET) * DN(TRK))/XI_T_(GAMMA) PP smeared}"
+"- header: {name: 1/N(JET) * DN(TRK))/XI_T_(GAMMA) PP smeared}"
+"- header: {name: 1/N(JET) * DN(TRK))/XI_T_(GAMMA) PP smeared}"
+"- header: {name: PBPB / PP}"
+"- header: {name: PBPB / PP}"
+"- header: {name: PBPB / PP}"
+"- header: {name: PBPB / PP}"
 );
 
 newLines3=(
@@ -43,10 +67,18 @@ newLines3=(
 "- header: {name: 1/N(JET) * DN(TRK))/DXI_(JET) PBPB}"
 "- header: {name: 1/N(JET) * DN(TRK))/DXI_(JET) PBPB}"
 "- header: {name: 1/N(JET) * DN(TRK))/DXI_(JET) PBPB}"
+"PLACEHOLDER"
+"PLACEHOLDER"
+"PLACEHOLDER"
+"PLACEHOLDER"
 "- header: {name: 1/N(JET) * DN(TRK))/XI_T_(GAMMA) PBPB}"
 "- header: {name: 1/N(JET) * DN(TRK))/XI_T_(GAMMA) PBPB}"
 "- header: {name: 1/N(JET) * DN(TRK))/XI_T_(GAMMA) PBPB}"
 "- header: {name: 1/N(JET) * DN(TRK))/XI_T_(GAMMA) PBPB}"
+"PLACEHOLDER"
+"PLACEHOLDER"
+"PLACEHOLDER"
+"PLACEHOLDER"
 );
 
 linesReaction=(
@@ -58,9 +90,25 @@ linesReaction=(
 "  - {name: RE, value: P P --> X PB PB --> X}"
 "  - {name: RE, value: P P --> X PB PB --> X}"
 "  - {name: RE, value: P P --> X PB PB --> X}"
+"  - {name: RE, value: P P --> X PB PB --> X}"
+"  - {name: RE, value: P P --> X PB PB --> X}"
+"  - {name: RE, value: P P --> X PB PB --> X}"
+"  - {name: RE, value: P P --> X PB PB --> X}"
+"  - {name: RE, value: P P --> X PB PB --> X}"
+"  - {name: RE, value: P P --> X PB PB --> X}"
+"  - {name: RE, value: P P --> X PB PB --> X}"
+"  - {name: RE, value: P P --> X PB PB --> X}"
 );
 
 linesCentrality=(
+"  - {name: CENTRALITY, value: '50 - 100%'}"
+"  - {name: CENTRALITY, value: '30 - 50%'}"
+"  - {name: CENTRALITY, value: '10 - 30%'}"
+"  - {name: CENTRALITY, value: '0 - 10%'}"
+"  - {name: CENTRALITY, value: '50 - 100%'}"
+"  - {name: CENTRALITY, value: '30 - 50%'}"
+"  - {name: CENTRALITY, value: '10 - 30%'}"
+"  - {name: CENTRALITY, value: '0 - 10%'}"
 "  - {name: CENTRALITY, value: '50 - 100%'}"
 "  - {name: CENTRALITY, value: '30 - 50%'}"
 "  - {name: CENTRALITY, value: '10 - 30%'}"
@@ -108,14 +156,16 @@ do
     newLine="dependent_variables:\n"${newLines2[i1]}"\n  qualifiers:\n"${linesReaction[i1]}"\n"${linesCentrality[i1]}"\n"$textKinematics"\n"$oldLine
     sed -i "0,/${oldLine}/! {0,/${oldLine}/ s;${oldLine};${newLine};}" $outputFile  # replace only the 2nd occurance in the file
 
-    # add header following 2nd "dependent variables"
-    oldLine=$oldLine2
-    newLine=${newLines3[i1]}"\n  qualifiers:\n"${linesReaction[i1]}"\n"${linesCentrality[i1]}"\n"$textKinematics"\n"$oldLine
-    outputFileTmp=$outputFile".tmp"
-    mv $outputFile $outputFileTmp
-    # replace only the 3rd occurance in the file
-    awk -v oldLineAWK="$oldLine" -v newLineAWK="$newLine" '/values/{count++;if(count==3){sub(oldLineAWK, newLineAWK)}}1' $outputFileTmp > $outputFile
-    rm $outputFileTmp
+    if [[ ! ${newLines3[i1]} = "PLACEHOLDER" ]]; then
+      # add header following 2nd "dependent variables"
+      oldLine=$oldLine2
+      newLine=${newLines3[i1]}"\n  qualifiers:\n"${linesReaction[i1]}"\n"${linesCentrality[i1]}"\n"$textKinematics"\n"$oldLine
+      outputFileTmp=$outputFile".tmp"
+      mv $outputFile $outputFileTmp
+      # replace only the 3rd occurance in the file
+      awk -v oldLineAWK="$oldLine" -v newLineAWK="$newLine" '/values/{count++;if(count==3){sub(oldLineAWK, newLineAWK)}}1' $outputFileTmp > $outputFile
+      rm $outputFileTmp
+    fi
 
     echo "modified : "$outputFile
 done
